Updated exception.py
Add module docstring: Adding a docstring at the beginning of the file can help explain what the file is for and how to use it. It can also provide information on the module's functionality and purpose.

Updated with PE8 conventions.

Avoid using print statements in class constructors they are  generally not recommended, as they can interfere with how the class is used by other parts of the code.